split up serialize module
